feat/useWageCheck 커스텀훅 분리
